update build system

+++ scripts/download.sh
@@ -6,17 +6,47 @@ mkdir -p $distfiles
 
 cd $distfiles
 
-wget -c "http://www.fftw.org/fftw-3.3.9.tar.gz"
+if [ -f "fftw-3.3.9.tar.gz" ] ; then
+    echo "fftw is downloaded"
+else
+    wget -c "http://www.fftw.org/fftw-3.3.9.tar.gz"
+fi
 
-wget -c "http://www.feynarts.de/cuba/Cuba-4.2.1.tar.gz"
+if [ -f "Cuba-4.2.1.tar.gz" ] ; then
+    echo "cuba is downloaded"
+else
+    wget -c "http://www.feynarts.de/cuba/Cuba-4.2.1.tar.gz"
+fi
 
-wget -c "https://zlib.net/zlib-1.2.11.tar.gz"
+if [ -f "zlib-1.2.11.tar.gz" ] ; then
+    echo "zlib is downloaded"
+else
+    wget -c "https://zlib.net/zlib-1.2.11.tar.gz"
+fi
 
-wget -c "https://gitlab.com/libeigen/eigen/-/archive/3.3.9/eigen-3.3.9.tar.bz2"
+if [ -f "eigen-3.3.9.tar.bz2" ] ; then
+    echo "eigen is downloaded"
+else
+    wget -c "https://gitlab.com/libeigen/eigen/-/archive/3.3.9/eigen-3.3.9.tar.bz2"
+fi
 
-wget -O c-lime.tar.gz -c "https://github.com/usqcd-software/c-lime/tarball/master"
+if [ -f "c-lime.tar.gz" ] ; then
+    echo "c-lime is downloaded"
+else
+    wget -O c-lime.tar.gz -c "https://github.com/usqcd-software/c-lime/tarball/master"
+fi
 
-wget -c "https://www.mpfr.org/mpfr-current/mpfr-4.1.0.tar.bz2"
+if [ -f "mpfr-4.1.0.tar.bz2" ] ; then
+    echo "mpfr is downloaded"
+else
+    wget -c "https://www.mpfr.org/mpfr-current/mpfr-4.1.0.tar.bz2"
+fi
+
+if [ -f "hdf5-1.12.1.tar.bz2" ] ; then
+    echo "hdf5 is downloaded"
+else
+    wget -c "https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.12/hdf5-1.12.1/src/hdf5-1.12.1.tar.bz2"
+fi
 
 if [ -d Grid ] ; then
     ( cd Grid ; git pull )
@@ -24,17 +54,40 @@ else
     git clone https://github.com/waterret/Grid.git
 fi
 
-(
-cd Grid
-pwd
-git clean -f
-./bootstrap.sh
-)
+if [ -e Grid/configure ] ; then
+    echo "Grid bootstrapped."
+else
+    ( cd Grid ; git clean -f ; ./bootstrap.sh )
+fi
 
 if [ -d gpt ] ; then
     ( cd gpt ; git pull )
 else
     git clone https://github.com/waterret/gpt.git
+fi
+
+if [ -d Grid-tblum ] ; then
+    ( cd Grid-tblum ; git pull )
+else
+    git clone https://github.com/tblum2/Grid.git Grid-tblum
+fi
+
+if [ -e Grid-tblum/configure ] ; then
+    echo "Grid-tblum bootstrapped."
+else
+    ( cd Grid-tblum ; git clean -f ; ./bootstrap.sh )
+fi
+
+if [ -d Hadrons-tblum ] ; then
+    ( cd Hadrons-tblum ; git pull )
+else
+    git clone https://github.com/tblum2/Hadrons.git Hadrons-tblum
+fi
+
+if [ -e Hadrons-tblum/configure ] ; then
+    echo "Hadrons-tblum bootstrapped."
+else
+    ( cd Hadrons-tblum ; git clean -f ; ./bootstrap.sh )
 fi
 
 cd $wd
